apps/server/requirements.txt
typing-extensions
langchain-openai
langchain-experimental
pandas
yfinance
praw
feedparser
stockstats
eodhd
langgraph
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
requests
tqdm
pytz
redis
chainlit
rich
questionary
langchain_anthropic
langchain-google-genai
fastapi
uvicorn
sse-starlette
sqlmodel
pyyaml
structlog
apscheduler
pydantic-settings
python-dotenv
httpx
numpy
psutil
